requirements.txt
streamlit
tensorflow
numpy
Pillow
python-dotenv